Feedback items can now be sorted by user option
I went ahead and added a comment property to the two feedback items that didn't have them and set it to an empty array in data.json. This is more in line with how data from a database would arrive when fetched. An item with 0 comments would not have the comments column/property removed completely in the database